requirements.txt
Flask==2.3.3
Werkzeug==3.0.1
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
et-xmlfile==1.1.0
numpy==1.24.4
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
wfastcgi==3.0.0
twilio==8.13.0
flask-login==0.6.3
bcrypt==4.1.2
flask-sqlalchemy==3.1.1
alembic==1.13.1 